llm_python/pyproject.toml
[project]
name = "llm-langchain"
version = "0.1.0"
description = ""
authors = [{name = "Russell Canfield", email = "rcanfield86@gmail.com"}]
readme = "README.md"
dependencies = [
  "langchain",
  "llama-cpp-python"
]